Fix address and signature for external C fn
In the first implementation I added the index of
RomDriver_PDS_Power_On_PLL to the base address.
This was incorrect, since the offset is 88 times the size of each
entry in the table - so we were pointing to RomDriver_DelayMS instead.
I also originally defined it as being fn(usize), as that was how it was
used in the SDK, but this is not how it was defined. It's now correctly
fn(usize) -> usize.