Fix PLL init & UART default clock, remove invalid PLL setting (#13)
* Fix address and signature for external C fn

In the first implementation I added the index of
RomDriver_PDS_Power_On_PLL to the base address.
This was incorrect, since the offset is 88 times the size of each
entry in the table - so we were pointing to RomDriver_DelayMS instead.
I also originally defined it as being fn(usize), as that was how it was
used in the SDK, but this is not how it was defined. It's now correctly
fn(usize) -> usize.

* Make default UART clock rate uart_clk

If the user doesn't choose a uart clock rate, it used to default to RC32
But, if the system clock is not RC32 this can make that clock impossible
to achieve. So, we set the default to the source clock rate instead,
which should always be possible.

* Remove invaild PLL setting

While the PLL outputs quite a few frequencies, only a few of them
are valid for sysclk.
80Mhz is not one of them.